requirements.txt
aiohappyeyeballs==2.3.6
aiohttp==3.10.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==24.2.0
autocommand==2.2.2
Automat==22.10.0
backoff==2.2.1
backports.tarfile==1.2.0
bcrypt==4.2.0
build==1.2.1
cachetools==5.4.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
coloredlogs==15.0.1
constantly==23.10.4
cryptography==43.0.0
cssselect==1.2.0
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
fastapi==0.112.1
filelock==3.15.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
google-auth==2.33.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.65.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
humanfriendly==10.0
hyperlink==21.0.0
idna==3.7
importlib_metadata==8.0.0
importlib_resources==6.4.2
incremental==24.7.2
inflect==7.3.1
itemadapter==0.9.0
itemloaders==1.3.1
jaraco.context==5.3.0
jaraco.functools==4.0.2
jaraco.text==4.0.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kubernetes==30.1.0
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.33
langchain-text-splitters==0.2.2
langsmith==0.1.99
llvmlite==0.43.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.19.0
openai==1.40.8
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
ordered-set==4.1.0
orjson==3.10.7
overrides==7.7.0
packaging==24.1
panda==0.3.1
parsel==1.9.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
posthog==3.5.0
Protego==0.3.1
protobuf==4.25.4
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyDispatcher==2.0.7
Pygments==2.18.0
pynndescent==0.5.13
pyOpenSSL==24.2.1
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
queuelib==1.7.0
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.4
scikit-learn==1.5.1
scipy==1.14.0
Scrapy==2.11.2
sentence-transformers==3.0.1
service-identity==24.1.0
setuptools==72.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.2
sympy==1.13.2
tenacity==8.5.0
threadpoolctl==3.5.0
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.0
triton==3.0.0
Twisted==24.7.0
typeguard==4.3.0
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
umap-learn==0.5.6
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
w3lib==2.2.1
watchfiles==0.23.0
websocket-client==1.8.0
websockets==12.0
wheel==0.44.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.20.0
zope.interface==7.0.1
